RFXCOM binding: Add support of RFY and fix Integer conversion of unitId coming from UI

+++ addons/binding/org.openhab.binding.rfxcom/src/main/java/org/openhab/binding/rfxcom/internal/messages/RFXComRfyMessage.java
@@ -219,7 +219,7 @@ public class RFXComRfyMessage extends RFXComBaseMessage {
 			throw new RFXComException("Invalid device id '" + deviceId + "'");
 		}
 
-		this.unitId = (byte) Short.parseShort(ids[0]);
+		this.unitId = Integer.parseInt(ids[0]);
 		this.unitCode = Byte.parseByte(ids[1]);
 	}
